ETL_AWS_ATHENA_DATABRICKS/requirements.txt
boto3
awswrangler
pandas
databricks-sql-connector
